requirements/env_futu.txt
pandas
pyarrow
openpyxl  # For downloading HKEX stock list
tqdm
requests
lxml
simplejson
click
python-dotenv
futu-api
